change to access 32-bit register

+++ memPhy2Virt/memPhy2Virt.h
@@ -4,8 +4,8 @@
 #define PAGE_SIZE 0x1000
 #define PAGE_SIZE_MASK 0xfffffffffffff000
 
-int showRegValue(unsigned long phyAddr, unsigned long len);
-int writeReg(unsigned long phyAddr, unsigned long value);
-int readReg(unsigned long phyAddr, unsigned long *value);
+int showRegValue(unsigned long phyAddr, unsigned int len);
+int writeReg(unsigned long phyAddr, unsigned int value);
+int readReg(unsigned long phyAddr, unsigned int *value);
 
 #endif
